pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telememo"
version = "0.1.0"
description = "Dump Telegram channel messages to SQLite"
requires-python = ">=3.10"
dependencies = [
    "telethon>=1.34.0",
    "peewee>=3.17.0",
    "click>=8.1.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "cryptg>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
telememo = "telememo.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["telememo*"]
